@@ -11,19 +11,19 @@
 	<url>http://maven.apache.org</url>
 
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>13</java.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<junit-platform-surefire-provider-version>1.0.1</junit-platform-surefire-provider-version>
 		<junit-launcher-version>1.0.1</junit-launcher-version>
-		<junit-jupiter-version>5.0.2</junit-jupiter-version>
+		<junit-jupiter-version>5.5.1</junit-jupiter-version>
 
 		<spring-version>5.0.4.RELEASE</spring-version>
 
 		<maven-compiler-plugin.version>3.7.0</maven-compiler-plugin.version>
-		<jacoco.version>0.8.1</jacoco.version>
+		<jacoco.version>0.8.4</jacoco.version>
 		<surefire-plugin-version>2.19.1</surefire-plugin-version>
 		<surefire-provider-version>1.0.1</surefire-provider-version>
-		<mockito-version>2.7.22</mockito-version>
+		<mockito-version>2.25.0</mockito-version>
 	</properties>
 
 	<dependencies>
@@ -48,12 +48,12 @@
 			<scope>test</scope>
 		</dependency>
 
-		<dependency>
+		<!-- dependency>
 			<groupId>org.junit.platform</groupId>
 			<artifactId>junit-platform-launcher</artifactId>
 			<version>${junit-launcher-version}</version>
 			<scope>test</scope>
-		</dependency>
+		</dependency-->
 
 		<!-- https://mvnrepository.com/artifact/org.mockito/mockito-core -->
 		<dependency>
